+++ src/agents/minimax.py
@@ -1,19 +1,39 @@
-from random import choice
 from game_logic.board import Board
 from agents.strategy import Strategy
 from game_logic.enums import *
 from copy import deepcopy
+import time
 
+# TODO: make a better order of moves to make alpha-beta better
+# TODO: Improve heuristic
 
 class Minimax(Strategy):
   """
   AI agent following a custom minimax with alpha-beta pruning policy.
   """
-
   color = PlayerColor.WHITE
 
   def utility(self, state, board):
-    # Cases where win or lose is defined
+    """
+    Heuristic function evaluating the current game state from the AI's perspective.
+
+    The heuristic value is defined as follows:
+
+    - If the game is a draw, the value is 0.
+    - If the AI wins, the value is 100.
+    - If the opponent wins, the value is -100.
+    - If the AI is white, the value is the difference between the number of bugs near the AI's queen and the number of bugs near the opponent's queen.
+    - If the AI is black, the value is the difference between the number of bugs near the AI's queen and the number of bugs near the opponent's queen.
+
+    This heuristic is very simple and does not take into account other factors such as the number of moves left, the number of pieces on the board, etc.
+
+    :param state: The current game state.
+    :type state: GameState
+    :param board: The current game board.
+    :type board: Board
+    :return: The heuristic value of the current game state.
+    :rtype: int
+    """
     if state == GameState.DRAW:
         return 0
     if self.color == PlayerColor.WHITE:
@@ -21,17 +41,66 @@ class Minimax(Strategy):
         return 100
       if state == GameState.BLACK_WINS:
         return -100
-      # Calculate other utility
-      return board._get_number_of_bugs_near_queen(PlayerColor.WHITE) - board._get_number_of_bugs_near_queen(PlayerColor.BLACK)
       
-    
-    if state == GameState.BLACK_WINS:
-      return 100
-    elif state == GameState.WHITE_WINS:
-      return -100
-    return board._get_number_of_bugs_near_queen(PlayerColor.BLACK) - board._get_number_of_bugs_near_queen(PlayerColor.WHITE)
+      return board._get_number_of_bugs_near_queen(PlayerColor.WHITE) - board._get_number_of_bugs_near_queen(PlayerColor.BLACK)
+    else:
+      if state == GameState.BLACK_WINS:
+        return 100
+      elif state == GameState.WHITE_WINS:
+        return -100
+      return board._get_number_of_bugs_near_queen(PlayerColor.BLACK) - board._get_number_of_bugs_near_queen(PlayerColor.WHITE)
+
+  def initiate_minimax(self, board):
+    """
+    Starts the minimax tree by checking all possible moves and returning the one with the highest utility.
+
+    If no actions are available, it returns the PASS command.
+
+    :param board: The current game board.
+    :type board: Board
+    :return: The best move according to the minimax algorithm.
+    :rtype: str
+    """
+    #start = time.perf_counter()
+    max_eval = -float('inf')
+    alpha = -float('inf')
+    todo_action = Command.PASS
+    # For every action available, play it and calculate the utility (recursively)
+    for action in board.valid_moves.split(";"):
+      bb = deepcopy(board)
+      bb.play(action)
+      eval = self.minmax(state=bb.state, board=board, depth=1, maximizing_player=False, alpha=alpha, beta=float('inf'))
+      if (eval > max_eval):
+        max_eval = eval
+        todo_action = action
+      alpha = max(alpha, max_eval)
+    #end = time.perf_counter()
+    #print(end-start)
+    return todo_action
 
   def minmax(self, state, board, depth, maximizing_player, alpha, beta):
+    """
+    Performs the minimax algorithm with alpha-beta pruning to evaluate and return the utility of a game state.
+
+    The function recursively explores all possible moves up to a specified depth to determine the best move for the
+    current player. It optimizes the search process by pruning branches that cannot produce a better outcome than
+    previously examined branches.
+
+    :param state: The current game state.
+    :type state: GameState
+    :param board: The current game board.
+    :type board: Board
+    :param depth: The current depth in the game tree.
+    :type depth: int
+    :param maximizing_player: Indicates whether the current player is maximizing or minimizing their score.
+    :type maximizing_player: bool
+    :param alpha: The best already explored option along the path to the root for the maximizer.
+    :type alpha: float
+    :param beta: The best already explored option along the path to the root for the minimizer.
+    :type beta: float
+    :return: The utility value of the current game state.
+    :rtype: int
+    """
     if state == GameState.DRAW or state == GameState.WHITE_WINS or state == GameState.BLACK_WINS or depth == self.depth_limit:
         return self.utility(state, board)
     
@@ -41,12 +110,11 @@ class Minimax(Strategy):
             bb = deepcopy(board)
             bb.play(action)
             eval = self.minmax(bb.state, bb, depth + 1, False, alpha, beta)
-            alpha = max(alpha, eval)
-            if beta <= alpha:
-                break  # Beta cut-off
-            if eval > max_eval and depth == 0:
-              self._cache = action
             max_eval = max(max_eval, eval)
+            alpha = max(alpha, max_eval)
+            # We cut all the remaining children
+            if beta <= alpha:
+                break  
         return max_eval
     else:
         min_eval = float('inf')
@@ -54,14 +122,27 @@ class Minimax(Strategy):
             bb = deepcopy(board)
             bb.play(action)
             eval = self.minmax(bb.state, bb, depth + 1, True, alpha, beta)
-            beta = min(beta, eval)
-            if beta <= alpha:
-                break  # Alpha cut-off
             min_eval = min(min_eval, eval)
+            beta = min(beta, min_eval)
+            # We cut all the remaining children
+            if beta <= alpha:
+                break  
         return min_eval
 
   def calculate_best_move(self, board: Board) -> str:
+    """
+    Calculates the best move for the current player given the state of the board.
+
+    If this is the first time the method is called, it will start the minimax tree and calculate the best move.
+    Otherwise, it simply returns the cached result.
+
+    :param board: The current game board.
+    :type board: Board
+    :return: The best move according to the minimax algorithm.
+    :rtype: str
+    """
+    
     if not self._cache:
       self.color = board.current_player_color
-      self.minmax(state=GameState.IN_PROGRESS, board=board, depth=0, maximizing_player=True, alpha=-float('inf'), beta=float('inf'))
+      self._cache = self.initiate_minimax(board)
     return self._cache